requirements.txt
pysocks==1.7.1
telethon==1.18.2
urlextract==1.2.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6
